Fix compiler warnings from optimized build.

+++ IncrementalSVDFastUpdate.C
@@ -184,7 +184,7 @@ IncrementalSVDFastUpdate::addNewSample(
       max_U_dim = d_num_samples;
    }
    if (fabs(checkOrthogonality(d_U)) >
-       std::numeric_limits<double>::epsilon()*max_U_dim) {
+       std::numeric_limits<double>::epsilon()*static_cast<double>(max_U_dim)) {
       reOrthogonalize(d_U);
    }
    if (fabs(checkOrthogonality(d_Up)) >
